Add U-Net AMP + one-shot hand-drawn training with end-to-end validation
train_unet: opt-in --amp / --amp-dtype (bf16 default; GradScaler only for
fp16). DiceLoss now reduces in fp32 so the sigmoid+spatial-sum stays safe
under autocast. AMP-off path is unchanged.

training/train_hand_drawn.{sh,bat}: one-shot hand-drawn pipeline (synthetic
+ auto-detected CGHD -> U-Net + YOLOv8). SMOKE=1 runs the whole chain fast
(tiny data -> 2 epochs -> inference -> .cir) to validate before the real run.

training/WINDOWS_SETUP.md: step-by-step RTX 4060 / Windows guide that
installs the CUDA torch wheel first to avoid the CPU-only build, with a
verify step.

+++ training/unet/losses.py
@@ -19,9 +19,12 @@ class DiceLoss(nn.Module):
         self.smooth = smooth
 
     def forward(self, logits: torch.Tensor, targets: torch.Tensor) -> torch.Tensor:
-        probs = torch.sigmoid(logits)
+        # .float() keeps the sigmoid and the large spatial sum in fp32 even under
+        # AMP autocast, where an fp16 reduction over ~590k pixels can lose
+        # precision or overflow. No effect when logits are already fp32.
+        probs = torch.sigmoid(logits.float())
         probs = probs.reshape(probs.shape[0], -1)
-        targets = targets.reshape(targets.shape[0], -1)
+        targets = targets.float().reshape(targets.shape[0], -1)
         intersection = (probs * targets).sum(dim=1)
         union = probs.sum(dim=1) + targets.sum(dim=1)
         dice = (2 * intersection + self.smooth) / (union + self.smooth)
